Implement Live Channel Updates with silent background refresh

+++ app/src/main/java/com/codesrahul/exclusivetv/UpdateManager.kt
@@ -26,7 +26,7 @@ import java.io.File
 
 class UpdateManager(
     private var context: Context,
-    private var versionCode: Long
+    private var versionCode: Int
 ) :
     ConfirmationFragment.ConfirmationListener {
